+++ src/main/java/prography/pingpong_game/common/exception/GlobalExceptionHandler.java
@@ -1,9 +1,10 @@
-package prography.pingpong_game.common;
+package prography.pingpong_game.common.exception;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+import prography.pingpong_game.common.ApiResponse;
 
 @RestControllerAdvice
 @Slf4j
